Fixed issue #2 - workaround defunkt Xwt(.Gtk3) TextLayout disposal.
I have reviewed memory usage growth with and without various UI parts and
identified that TextLayout is the (most) notable cause of the issue.

There were many TextLayout allocations in the old code, which was definitely
app code issue. Those objects are disposables that we should dispose by
ourselves.

The expected fix was (obviously) to dispose those objects.

What happened after that? They don't seem to be disposed at all(!)

I once wrote an XWT helper method `DrawingHelper.DrawText()` which draws
text in a method call with a bunch of arguments. I have replaced every
`new TextLayout()` allocations with this helper method, and then
tried to reuse one TextLayout object using spinlock (which does not
seem to actually lock the thread because of zero-async UI rendering in
the app). And the huge memory consumption is gone.

The corresponding code is in `#if false ... #else ... #endif` block.