various reference fixes and removal of extra dependency code to get it working on Mac and WPF.
Xwt.XamMac needs specific build, so remove it from desktop project.
And in desktop proj, use Xwt.Gtk for that purpose instead.
For XamMac proj, there seems to be extraneous reference to Gtk in Xmdsp.Xwt.

There was another issue for WPF - there was a duplicate menu item addition
which somehow did not break on Gtk but did for WPF and XamMac.